# Improve string.Equals OrdinalIgnoreCase performance for matching chars