@danmosemsft 2017-02-17
LGTM. @stephentoub any concerns?