@stephentoub 2017-02-17
@bbowyersmyth, thanks for sharing the perf results.  All of the tables appear to be for cases where the strings are equal (either with or without the same case); what about cases where they're not, e.g. where the first char differs, where a middle char differs, etc.?