@bbowyersmyth 2017-02-17
@stephentoub They had almost the same results so I left them off. eg First char difference is 1.8ns old vs 1.4ns new. Differences at longer indexes have the delta added to the equals cost up to that point.