@danmosemsft 2017-02-19
@bbowyersmyth if you rebase we can merge this.