@bbowyersmyth 2017-02-21
@danmosemsft rebased